Add info about remark-math and using $ characters
Hi,

Thanks for this! I was moving over my site to mdsvex and this plugin saved a couple of my posts :) It look me a little while to figure out why it wasn't working until I realized I needed to use `remark-math` and to write to the formulas with `$` not `<div class="math-display>`. These are a couple of clarifications that I hope will make it easier for others to use this plugin